add model entry

+++ handler/upload.js
@@ -4,6 +4,7 @@ const path = require('path');
 const hostname = process.env.NODE_ENV !== 'production' ?
     'localhost' : '35.188.36.119';
 const fs = require('fs');
+const { diffieHellman } = require('crypto');
 let uploadfiles = {
     files: [],
 };
@@ -37,11 +38,15 @@ const addFileUploadHandler = (req, res) => {
             throw Error(req.rval);
         }
         const {filename, mimetype} = req.file;
+        const {model} = req.query;
+        console.log(model);
+        if (!model) { throw Error('model not found'); }
         // add new entry
         // TODO: Change localhost with external IP
         const newFile = {
             filename: filename,
             mimetype: mimetype,
+            model: model,
             url: 'http://' + hostname + ':5000' + '/download/' + filename,
         };
         uploadfiles.files.push(newFile);
@@ -50,6 +55,7 @@ const addFileUploadHandler = (req, res) => {
         return res.status(200).json({
             status: 'success',
             filename: filename,
+            model: model,
             url: 'http://' + hostname + ':5000' + '/download/' + filename,
         });
     } catch (e) {
@@ -72,6 +78,7 @@ const deleteFileUploadHandler = (req, res) => {
             if (err) throw Error('files entry already cleared');
 
             for (const file of files) {
+                if (file === '.gitkeep') continue;
                 fs.unlink(path.join(directory, file), (err) => {
                     if (err) throw Error('files entry already cleared');
                 });
